+++ lib/types.ts
@@ -9,6 +9,14 @@ export interface SkillSettings {
     userId : string;
     deviceId : string;
     locale : string;
+    interfaces? : InterfaceSettings;
+}
+
+export interface InterfaceSettings {
+    display? : boolean;
+    audio? : boolean;
+    video? : boolean;
+    apl? : boolean;
 }
 
 export interface SequenceItem {