+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 ## current master
+* add supportedInterfaces to request builder
 
 ## 2.0.1
 * adding more supported types to attributes check